refactor: interface -> type으로 컨벤션 통일
