+++ frontend/src/components/@shared/Modal/BottomSheet.tsx
@@ -5,13 +5,13 @@ import { onMountFromBottomModal, unMountToBottomModal } from '../../../styles/An
 
 type BottomSheetModalSize = 'large' | 'medium' | 'small';
 
-interface IBottomSheetLayout {
+type BottomSheetProps = {
   children: ReactNode;
   size?: BottomSheetModalSize;
-}
+};
 
 export default forwardRef(
-  ({ children, size }: IBottomSheetLayout, ref: LegacyRef<HTMLDivElement>) => {
+  ({ children, size }: BottomSheetProps, ref: LegacyRef<HTMLDivElement>) => {
     return (
       <S.Layout ref={ref} size={size}>
         <S.Inner>{children}</S.Inner>
