+++ frontend/src/types/user.ts
@@ -1,6 +1,6 @@
-export interface UserProfile {
+export type UserProfile = {
   id: number;
   name: string;
   email: string;
   imageUrl: string;
-}
+};
